StreamMetrics/requirements.txt
confluent-kafka
pyspark
google-cloud-bigquery
